% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{Test Assumptions}
\usage{
assumptions(question, data)
}
\arguments{
\item{question}{the question number; an integer between 1 and 3 inclusive}

\item{data}{the data input; must be of class "data.frame"}
}
\value{
the correct plots or graph to test assumptions
}
\description{
Test Assumptions
}
