% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion.R
\name{conclusion}
\alias{conclusion}
\title{Report test conclusion}
\usage{
conclusion(question, data)
}
\arguments{
\item{question}{the question number; an integer between 1 and 3 inclusive}

\item{data}{the data input; must be of class "data.frame"}
}
\value{
the conclusion of the statistical test
}
\description{
Report test conclusion
}
