% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision}
\alias{decision}
\title{Report test decision}
\usage{
decision(question, data)
}
\arguments{
\item{question}{the question number; an integer between 1 and 3 inclusive}

\item{data}{the data input; must be of class "data.frame"}
}
\value{
the decision of the statistical test
}
\description{
Report test decision
}
