% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Report test summary}
\usage{
fit(data, question, output = "LaTeX")
}
\arguments{
\item{data}{the data input; must be of class "data.frame"}

\item{question}{the question number; must be an integer between 1 and 3 inclusive}

\item{output}{the desired type of output; must be either "LaTeX" or "R"}
}
\value{
a statistical test summary to help answer the question
}
\description{
Report test summary
}
\examples{
fit(proj, 1, output = "R")
}
