% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytest.R
\name{mytest}
\alias{mytest}
\title{Conduct a complete statistical test}
\usage{
mytest(data, question)
}
\arguments{
\item{data}{the data input; must be of class "data.frame"}

\item{question}{the question number; must be an integer between 1 and 3 inclusive}
}
\value{
a complete statistical test
}
\description{
Conduct a complete statistical test
}
\examples{
mytest(pro, 1)
}
