% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mytest}
\alias{print.mytest}
\title{print.mytest}
\usage{
\method{print}{mytest}(x, ...)
}
\arguments{
\item{x}{an object of class "mytest"}

\item{...}{further arguments passed to or from other methods}
}
\value{
a nicely formatted print
}
\description{
print.mytest
}
\examples{
print(mytest(pro, 1))
}
