% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\docType{data}
\name{proj}
\alias{proj}
\title{Project data}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
\item{ID}{a unique ID for each person}
\item{gender}{the gender of the person}
\item{height}{the height of the person, in cm}
\item{weight}{the weight of the person, in kg}
\item{phys}{the level of physical activity of the person, either "None", "Moderate", or "Intense"}
}
}
\usage{
proj
}
\description{
Data of a sample of 1000 people
}
\keyword{datasets}
