% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision}
\alias{decision}
\title{Report test decision}
\usage{
decision(data, question, output = "LaTeX")
}
\arguments{
\item{data}{the data input; must be of class "data.frame"}

\item{question}{the question number; must be an integer between 1 and 3 inclusive}

\item{output}{the desired type of output; must be either "LaTeX" or "R"}
}
\value{
the decision of the statistical test
}
\description{
Report test decision
}
\examples{
decision(pro, 1, output = "R")
}
