% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Report test summary}
\usage{
fit(question, data)
}
\arguments{
\item{question}{the question number; an integer between 1 and 3 inclusive}

\item{data}{the data input; must be of class "data.frame"}
}
\value{
important values for the statistical test to answer the question
}
\description{
Report test summary
}
