% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis}
\alias{hypothesis}
\title{Hypothesis}
\usage{
hypothesis(question)
}
\arguments{
\item{question}{the question number; an integer between 1 and 3 inclusive}
}
\value{
a character string with the correct hypothesis for LaTeX to read
}
\description{
Hypothesis
}
