% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis}
\alias{hypothesis}
\title{Report test hypothesis}
\usage{
hypothesis(data, question, output = "LaTeX")
}
\arguments{
\item{data}{unnecessary argument used for consistency; must be of class "data.frame"}

\item{question}{the question number; must be an integer between 1 and 3 inclusive}

\item{output}{the desired type of output; must be either "LaTeX" or "R"}
}
\value{
a character string with the correct hypothesis
}
\description{
Report test hypothesis
}
\examples{
hypothesis(pro, 1, output = "R")
}
